added max_examples to brownie config for hypothesis and moved market tests into single file structure + dirs per concern
