+++ tests/markets/test_config.py
